Split spark 3.1 code from the spark3 project
