+++ integration/spark/build.gradle
@@ -55,6 +55,7 @@ dependencies {
     implementation(project(path: ":shared", configuration: "shadow"))
     implementation(project(path: ":spark2"))
     implementation(project(path: ":spark3"))
+    implementation(project(path: ":spark31"))
     implementation(project(path: ":spark32"))
     implementation(project(path: ":spark33"))
     implementation(project(path: ":spark34"))
@@ -223,6 +224,7 @@ shadowJar {
         exclude (project(":shared"))
         exclude (project(":spark2"))
         exclude (project(":spark3"))
+        exclude (project(":spark31"))
         exclude (project(":spark32"))
         exclude (project(":spark33"))
         exclude (project(":spark34"))
