+++ integration/spark/settings.gradle
@@ -1,5 +1,6 @@
 include 'spark2'
 include 'spark3'
+include 'spark31'
 include 'spark32'
 include 'spark33'
 include 'spark34'
