+++ integration/spark/spark31/src/test/resources/io/openlineage/spark/agent/version.properties
@@ -1,0 +1,1 @@
+version 0.11.0-SNAPSHOT
